% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_dives.R
\name{cum_dives}
\alias{cum_dives}
\title{Cumulative dives plot}
\usage{
cum_dives(path = getwd())
}
\arguments{
\item{path}{directory location where the files will be read from or written into}
}
\value{
Returns a plot.
}
\description{
Creates a plot with the number of dives along time
}
\details{
This function creates some visualizations of the number of dives along time including a density plot and a fitted model
}
\author{
Ruben Perez Perez
}
