% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divesite_depths.R
\name{divesite_depths}
\alias{divesite_depths}
\title{Depth variation per divesite}
\usage{
divesite_depths(toofew = 1)
}
\arguments{
\item{toofew}{number of logged dives per divesite under which a divesite is excluded
from the boxplot for having too few dives for an appropriate interpretation of the depth variation}
}
\value{
Returns a boxplot.
}
\description{
Creates a plot with the variation of depths of the logged dives
}
\details{
This function creates a boxplot to visualize the depth variation of the logged dives, offering an overview of the depth conditions
and opportunities of each divesite
}
\author{
Ruben Perez Perez
}
