% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divetypes_platform.R
\name{divetypes_platform}
\alias{divetypes_platform}
\title{Number of dives per dive platform type}
\usage{
divetypes_platform(path = getwd())
}
\value{
Returns a barplot.
}
\description{
Creates a barplot with the number of dives per region and country and platform type
}
\details{
This function creates a barplot to visualize the number of dives per region and platform type "poolDive", "boatDive" or "shoreDive"
}
\author{
Ruben Perez Perez
}
