% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divewatch.R
\name{divewatch}
\alias{divewatch}
\title{Creates an overview of the logbook}
\usage{
divewatch(data = NA, email = NA)
}
\arguments{
\item{data}{The URL ID of the database in google sheets, if no URL is provided a mock dataset generated
by the test_data function will be read}

\item{email}{email account that will be used to access the Logbook in Google Spreadsheet}
}
\value{
Creates an overview.html file with an overview of the diving logbook and creates a Fig folder with the images from the overview
}
\description{
Creates an .html file with an overview of the diving logbook loaded
}
\details{
This function uses all the other functions of the divewatchr R package to create a visual overview
of the diving logbook loaded using plots and maps in order to to explore it
}
\examples{
divewatch()
\dontrun{
divewatch("1PpXTVS8LdzbvwLHyAAhR2MdT9Iwdy-hiqJknUzF7Yqo")
}
}
\author{
Ruben Perez Perez
}
