% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divewatchr.R
\docType{package}
\name{divewatchr}
\alias{divewatchr}
\title{divewatchr: A package for visualizing and exploring a SCUBA Diving Logbook}
\description{
This package allows for a thorough visualization of a diving logbook,
including spatio-temporal and categorical plots and maps.
The divewatchr package provides functions to read data from a diving logbook,
create informative maps, plots and data frames about the dataset and to create
descriptive reports about the diving logbook
}
