% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data}
\alias{example_data}
\title{Creates a example dataset}
\usage{
example_data(path = getwd())
}
\arguments{
\item{path}{directory location where the files will be read from or written into}
}
\value{
4 csv files with example data.
}
\description{
Creates a example dataset with the basic data to perform the visualizations of the package
}
\details{
This function creates an example dataset with the structure
needed to perform the visualization functions from the divewatchr package.
It can be used to run the other functions and to show the user how the data must look like.
The output dataset must be copypasted in a google sheets spreadsheet and the user
needs to make sure that the data is formatted as character/text
}
\examples{
example_data()
}
\author{
Ruben Perez Perez
}
