% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_auth.R
\name{google_auth}
\alias{google_auth}
\title{Uses the edit rights to a Google sheets if they are stored by the user}
\usage{
google_auth(email = NA)
}
\arguments{
\item{email}{email account that will be used to access the Logbook in Google Spreadsheet}
}
\value{
It provides access to a given Google sheet if there is a token stored.
}
\description{
Stores the authentication token in a file and uses it to access the Google sheet
}
\details{
This function allows Github Action and a users that has a token stored to access a Google sheet to
retrieve the data from a given google sheet
}
\examples{
divewatch()
\dontrun{
google_auth()
}
}
\author{
Ruben Perez Perez
}
