% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loading data}
\usage{
load_data(data)
}
\arguments{
\item{data}{The URL ID of the database in google spreadhsets.}
}
\value{
Reads one R list and the 2 dataframes within it into the R environment.
}
\description{
Loads logbook data from the data repository into the R session
}
\details{
This function reads, updates and prepares the data from the data repository
so it can be uses by the rest of the functions
}
\examples{
load_data("1qO7_0K1R-4i_MSgtT3zAYbZfxBmPMgQWAi7OmWmb1-M")
}
\author{
Ruben Perez Perez
}
