% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_highlights.R
\name{logbook_highlights}
\alias{logbook_highlights}
\title{Logbook highlights}
\usage{
logbook_highlights(path = getwd())
}
\arguments{
\item{path}{directory location where the files will be read from or written into}
}
\value{
Returns a data frame.
}
\description{
Creates a data frame with the highlights of the logbook
}
\details{
This function creates a data frame to have an overview of some highlights of the diving logbook
such as number of recorded dives, maximum bottom time or maximum depth
}
\author{
Ruben Perez Perez
}
