% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{logbook_map}
\alias{logbook_map}
\title{Logged dives map}
\usage{
logbook_map()
}
\value{
Returns a map.
}
\description{
Creates a leaflet map of the logged dives
}
\details{
This function creates a leaflet map to visualize the geographical coverage of the logbook together with some information on each dive such as
Bottom Time, Max Depth, Date and Divesite
}
\author{
Ruben Perez Perez
}
