% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_depths.R
\name{logged_depths}
\alias{logged_depths}
\title{Logged dive max depths}
\usage{
logged_depths(path = getwd())
}
\value{
Returns a boxplot.
}
\description{
Creates a scatterplot with the logged maximum depths through time
}
\details{
This function creates a scatterplot to visualize the maximum depths logged through time, offering a nice overview
of the diving experience regarding depth
}
\author{
Ruben Perez Perez
}
