% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scuba_cert.R
\docType{data}
\name{scuba_cert}
\alias{scuba_cert}
\title{Example of the certifications table}
\format{
A data frame with 4 rows and 5 variables:
\describe{
\item{diverID}{the diver identifier linked to the certification}
\item{certificationName}{the name of the certification}
\item{certificationAgency}{the name or acronym of the agency that issued the certification}
\item{certificationDate}{the date when the certification was issued, in ISO8601}
\item{instructorID}{the diver identifier of the instructor who taught the course and issued the certification}
}
}
\usage{
scuba_cert
}
\description{
A dataset containing an example of the certifications table, this table contains
an example of what the column names must be and how the content must look like.
}
\keyword{datasets}
