% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{test_data}
\alias{test_data}
\title{Creates a fake dataset}
\usage{
test_data(path = getwd())
}
\arguments{
\item{path}{directory location where the files will be read from or written into}
}
\value{
2 csv files with fake data.
}
\description{
Creates a fake dataset with the basic data to perform the visualizations of the package
}
\details{
This function creates a model dataset with the structure
needed to perform the visualization functions from the divewatchr package.
It can be used to test the other functions and to show the user how the data must look like.
The output dataset must be copypasted in a google sheets spreadsheet and the user
needs to make sure that the data is formatted as character/text
}
\examples{
test_data()
}
\author{
Ruben Perez Perez
}
