% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Updating data}
\usage{
update_data(
  data = "1qO7_0K1R-4i_MSgtT3zAYbZfxBmPMgQWAi7OmWmb1-M",
 
    path = "C:/Users/rubenp/Desktop/Huiswerk 2020/Logbook Scuba Ruben Perez/divewatchr/data/"
)
}
\arguments{
\item{data}{The URL ID of the database in google spreadhsets.}

\item{path}{The location where the formatted database will be stored}
}
\value{
Reads one R list and the 2 dataframes within it into the R environment.
}
\description{
Updates and stores the .rda and .RData data files in the chosen location
}
\details{
This function reads, updates and prepares the data from the data repository,
it also locates it in the chosen repository so it can be pulled by the rest of the functions as
a data example
}
\examples{
load_data()
}
\author{
Ruben Perez Perez
}
