% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divecount_reg.R
\name{divecount_reg}
\alias{divecount_reg}
\title{Number of dives per region plot}
\usage{
divecount_reg()
}
\value{
Returns a barplot.
}
\description{
Creates a plot with the number of dives per region and country
}
\details{
This function creates a barplot to visualize the number of dives per region
}
\author{
Ruben Perez Perez
}
