% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divesite_depths.R
\name{divesite_depths}
\alias{divesite_depths}
\title{Depth variation per divesite}
\usage{
divesite_depths()
}
\value{
Returns a boxplot.
}
\description{
Creates a plot with the variation of depths of the logged dives
}
\details{
This function creates a boxplot to visualize the depth variation of the logged dives, offering an overview of the depth conditions
and opportunities of each divesite
}
\author{
Ruben Perez Perez
}
