% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divetypes.R
\name{divetypes}
\alias{divetypes}
\title{Number of dives per dive type}
\usage{
divetypes(n.other = 4, path = getwd())
}
\arguments{
\item{n.other}{number of dives under which a diveType is placed into the "Other" category.}
}
\value{
Returns a barplot.
}
\description{
Creates a barplot with the number of dives types
}
\details{
This function creates a barplot to visualize the number of dives types "Course Leading", "Course Assistance", etc.
}
\author{
Ruben Perez Perez
}
