% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepares the data}
\usage{
prep_data(
  data = "1PpXTVS8LdzbvwLHyAAhR2MdT9Iwdy-hiqJknUzF7Yqo",
  path = getwd()
)
}
\arguments{
\item{data}{The URL ID of the database in google sheets.}
}
\value{
Reads one R list and the 2 dataframes within it into the R environment.
}
\description{
Loads and prepares the logbook data from the data repository into the R session
}
\details{
This function reads, prepares and locates in the chosen repository the data from the database
so it can be used by the rest of the functions
}
\examples{
prep_data("1PpXTVS8LdzbvwLHyAAhR2MdT9Iwdy-hiqJknUzF7Yqo")
}
\author{
Ruben Perez Perez
}
