% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepares the data}
\usage{
prep_data(data = NA, path = getwd())
}
\arguments{
\item{data}{The URL ID of the database in google sheets, if no URL is provided a mock dataset generated
by the test_data function will be read}

\item{path}{directory location where the files will be read from or written into}
}
\value{
Reads one R list and the 2 dataframes within it into the R environment.
}
\description{
Loads and prepares the logbook data from the data repository into the R session
}
\details{
This function reads, prepares and locates in the chosen repository the data from the database
so it can be used by the rest of the functions
}
\examples{
prep_data()
}
\author{
Ruben Perez Perez
}
